Lots of changes... where to begin (see description)
Changed prediction model slightly

Increased floating type for loaded data (this will allow for better training and accuracy)

Added two ways to use the data. The old ascii way (convert amino acid to ascii number) or using a dict where each amino acid has a value

Added a way to stop the training much earlier (press ctrl+c). This has slowed performance slightly (~10%) but I believe that it is worth it. This will also make the script restartable if it has crashed.

Added a ROC curve and classification if the model is good or not.

Added calculations of metrics (acc score, precision core, sensitivity score etc...) At the moment, they are just printed to the terminal.

Added so that some parameters are shown in a table after training. I will add so that it is instead appended to a file.

Started working on a ways to visualize the model. Two methods have been used: Graphviz, which creates a file with an overview of the model and networkx to view the connections. The connections are not implemented yet, but the nodes are.

Added a prototype for a GAN! This can be trained to generate "false" signal peptides. I have not had any luck with it yet, since the generator model way outclassed by the discriminator (exactly the same as the classifier).