Many changes
Change so that data augmentation uses bool instead of a string

Made so that tensorflow does not allocate full memory on gpu. This allows for multiple processes and other programs to function on a computer that is training

Fixed shuffling of data when the shapes did not match

Made so that vectorization data takes less room, by changing it from float64 to float32 (could actually be int instead when I think about it). This would allow much larger datasets to be loaded.