Many changes
Changed variable name "cutoff" to "seq_length"

Made labels floats and not ints

Fixed data augmentation bug when using vectorization

Fixed so that a bug where all the labels were flipped is fixed.

Made the vectorize_to_seq using sampling instead of rounding. (i.e 0.4 is probability that it is amino acids, not rounding it to 0)